backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.6
async-timeout==4.0.2
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
deta==1.1.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.1.6
MarkupSafe==2.1.1
multidict==6.0.2
Pillow==9.1.0
psycopg2-binary
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
twilio==7.9.0
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
